chore: rename streaminglogs to replicalogs
